requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
xgboost==2.0.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
python-dotenv==1.0.1
anthropic==0.40.0
reportlab==4.1.0
requests==2.31.0
matplotlib==3.8.3
seaborn==0.13.2
joblib==1.3.2
pydantic==2.6.1
pytest==8.0.2
google-generativeai>=0.8.6